# back-furniture/src/api/swagger.yaml
openapi: 3.0.0
info:
  title: Furniture Store
  description: This is Furniture Shop Backend 
  version: 0.1.9
servers:
  - url: /api/v1
security:
  - BearerAuth: []
securityDefinitions:
  BearerAuth:
    type: apiKey
    name: Authorization
    in: header
    description: Enter your Bearer token here
consumes:
  - application/json 
paths:
...
  /product:
    get:
      tags:
        - "Product"
      x-swagger-router-controller: product
      operationId: getProduct
      summary: Returns a list of products.
      parameters:
        - name: currentPage
          in: query
          description: currentPage
          required: false
          schema:
            type: integer
            minimum: 1
        - name: pageSize
          in: query
          description: pageSize
          required: false
          schema:
            type: integer
            minimum: 1
        - name: sort
          in: query
          description: "sort column"
          required: false
          schema:
            type: integer
        - name: sortDirection
          in: query
          description: "sort i.e asc desc"
          required: false
          schema:
            type: integer
            enum: 
              - asc
              - desc                                   
      responses:
        '200':    # status code
          description: Default Response
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/Products"
    post: 
      x-swagger-router-controller: product
      operationId: addProduct
      summary: Register a new product.
      tags:
        - "Product"      
      description: Register a new productHome
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                productName:
                  type: string
                  example: Door
                image:
                  type: string
                  example: sumit
                description:
                  type: string
                  example: good quality
                category:
                  type: string
                  example: Doors
                price:
                  type: number
                  example: 233
                homecategory:
                  type: string
                  example: bed
                subcategory:
                  type: string
                  example: good quality
                availability:
                  type: boolean
                  example: true
                quantity:
                  type: number
                  example: 2
                rating:
                  type: number
                  example: 4                                   
      responses: 
        '200':
          description: Created 
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"      
    delete:
      tags:
        - "Product"
      x-swagger-router-controller: product
      operationId: deleteAllProduct
      summary: Delete All Available product.
      responses:
        '204':    # status code
          description: Default Response
  /product/{productId}:
    get:
      tags:
        - "Product"
      x-swagger-router-controller: product
      operationId: getProductById
      summary: Returns  product.
      parameters:
        - name: productId
          in: path
          description: productId
          required: false
          schema:
            type: string
      responses:
        '200':    # status code
          description: Default Response
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/Product"
    patch: 
      x-swagger-router-controller: product
      operationId: updateProductById
      summary: update in a product.
      tags:
        - "Product"      
      description: Register a new productHome
      parameters:
        - name: productId
          in: path
          description: productId
          required: false
          schema:
            type: string      
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                productName:
                  type: string
                  example: Door
                image:
                  type: string
                  example: sumit
                description:
                  type: string
                  example: good quality
                category:
                  type: string
                  example: Doors
                price:
                  type: number
                  example: 1
                homecategory:
                  type: string
                  example: bed
                subcategory:
                  type: string
                  example: good quality
                availability:
                  type: boolean
                  example: true
                quantity:
                  type: number
                  example: 3
                rating:
                  type: number
                  example: 4
      responses: 
        '200':
          description: Created 
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"      
    delete:
      tags:
        - "Product"
      x-swagger-router-controller: product
      operationId: deleteProductById
      summary: Delete All Available product.
      parameters:
        - name: productId
          in: path
          description: productId
          required: false
          schema:
            type: string        
      responses:
        '204':    # status code
          description: Default Response   
...
  /cart:
    get:
      tags:
        - "cart"
      x-swagger-router-controller: cart
      operationId: getProductInCart
      summary: Returns a list of products.
      parameters:
        - name: currentPage
          in: query
          description: currentPage
          required: false
          schema:
            type: integer
            minimum: 1
        - name: pageSize
          in: query
          description: pageSize
          required: false
          schema:
            type: integer
            minimum: 1
        - name: sort
          in: query
          description: "sort column"
          required: false
          schema:
            type: integer
        - name: sortDirection
          in: query
          description: "sort i.e asc desc"
          required: false
          schema:
            type: integer
            enum: 
              - asc
              - desc                                   
      responses:
        '200':    # status code
          description: Default Response
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/carts"
    post: 
      x-swagger-router-controller: cart
      operationId: addProductInCart
      summary: Register a new productInCart.
      tags:
        - "cart"      
      description: Register a new productInCart
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                productName:
                  type: string
                  example: Door
                image:
                  type: string
                  example: xyz
                price:
                  type: number
                  example: 233
                quantitySelected:
                  type: number
                  example: 2
                brand:
                  type: string
                  example: sara   
                productId:
                  type: string
                  example: sfsfs
                productDetailId:
                  type: string
                  example: dadsfs   
                userId:
                  type: string
                  example: dadsfs                                                                   
      responses: 
        '200':
          description: Created 
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/cart"      
    delete:
      tags:
        - "cart"
      x-swagger-router-controller: cart
      operationId: deleteAllProductInCart
      summary: Delete All Available product.
      responses:
        '204':    # status code
          description: Default Response
  /cart/{cartId}:
    get:
      tags:
        - "cart"
      x-swagger-router-controller: cart
      operationId: getProductInCartById
      summary: Returns  product.
      parameters:
        - name: cartId
          in: path
          description: cartId
          required: false
          schema:
            type: string
      responses:
        '200':    # status code
          description: Default Response
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/cart"
    patch: 
      x-swagger-router-controller: cart
      operationId: updateProductInCartById
      summary: update in a product.
      tags:
        - "cart"      
      description: Update a new productInCart
      parameters:
        - name: cartId
          in: path
          description: cartId
          required: false
          schema:
            type: string      
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                productName:
                  type: string
                  example: Door
                image:
                  type: string
                  example: xyz
                price:
                  type: number
                  example: 233
                quantitySelected:
                  type: number
                  example: 2
                brand:
                  type: string
                  example: sara  
      responses: 
        '200':
          description: Created 
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/cart"      
    delete:
      tags:
        - "cart"
      x-swagger-router-controller: cart
      operationId: deleteProductInCartById
      summary: Delete All Available productInCart.
      parameters:
        - name: cartId
          in: path
          description: cartId
          required: false
          schema:
            type: string        
      responses:
        '204':    # status code
          description: Default Response                         
components:
  securitySchemes:
    BearerAuth:
      type: apiKey
      name: Authorization
      in: header
      description: Enter your Bearer token here
  schemas:
...
    Product:
      type: object
      properties:
                productName:
                  type: string
                  example: Door
                image:
                  type: string
                  example: sumit
                description:
                  type: string
                  example: good quality
                category:
                  type: string
                  example: Doors
                price:
                  type: number
                  example: Door
                homecategory:
                  type: string
                  example: bed
                subcategory:
                  type: string
                  example: good quality
                availability:
                  type: string
                  example: Doors
                quantity:
                  type: number
                  example: good quality
                rating:
                  type: string
                  example: Doors 
    Products:
      type: array
      items:
        $ref: "#/components/schemas/Product"  
...
    cart:
      type: object
      properties:
                productName:
                  type: string
                  example: Door
                image:
                  type: string
                  example: xyz
                price:
                  type: number
                  example: 233
                quantitySelected:
                  type: number
                  example: 2
                brand:
                  type: string
                  example: sara  
                productId:
                  type: string
                  example: sfsfs
                productDetailId:
                  type: string
                  example: dadsfs                    
    carts:
      type: array
      items:
        $ref: "#/components/schemas/cart"                 